refactor: preview&apply use the same cmd pattern
